requirements.txt
discord.py[voice]
requests
psycopg2
sentry-sdk
pycountry
typing
dblpy
pyddblapi
psutil
pbwrap
lavalink
aiodns
cchardet
PyMoe